+++ board-back/src/main/java/com/zoo/boardback/domain/searchLog/dao/SearchLogRepository.java
@@ -3,6 +3,6 @@ package com.zoo.boardback.domain.searchLog.dao;
 import com.zoo.boardback.domain.searchLog.entity.SearchLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SearchLogRepository extends JpaRepository<SearchLog, Integer> {
+public interface SearchLogRepository extends JpaRepository<SearchLog, Long> {
 
 }
